+++ cppdecoder/sample/include/Frame.h
@@ -13,12 +13,12 @@ class Frame {
  public:
   explicit Frame(const std::vector<float>& values);
   std::vector<float>& getFeatures() { return features; }
-  size_t getDim() const { return dim; }
+  uint32_t getDim() const { return dim; }
   void show_content();
 
  private:
   std::vector<float> features;
-  size_t dim;
+  uint32_t dim;
 };
 
 #endif  // FRAME_H_
